new update
are separated from component stylesnew update

are separated from component styles
