+++ renta_facil_colima/.eslintrc.js
@@ -1,4 +1,9 @@
 module.exports = {
   root: false,
   extends: '@react-native',
+  parser: '@babel/eslint-parser', // Utilizar el analizador de Babel para archivos JSX
+  rules: {
+    // Agregar reglas personalizadas si es necesario
+  },
+  requireConfigFile: false, // Desactivar la verificación de archivos de configuración de Babel
 };
